pyproject.toml
[tool.poetry]
name = "agentic-rag-app"
version = "0.1.0"
description = "Agentic Analyze Application with Streamlit"
authors = ["yoichiro abe <yoichiro.abe@avanade.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.28.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
streamlit-authenticator = "^0.4.2"
extra-streamlit-components = "^0.1.80"
autogen-agentchat = "^0.6.2"
autogen-ext = {extras = ["azure", "docker", "openai"], version = "^0.6.2"}
python-dotenv = "^1.1.1"
pulp = "^3.2.1"
scikit-learn = "^1.7.0"
lightgbm = "^4.6.0"
ortools = "9.12.4544"
duckduckgo-search = "^8.0.4"
ruff = "^0.12.1"
seaborn = "^0.13.2"
azure-storage-blob = "^12.25.1"
matplotlib = "^3.10.3"
yfinance = "^0.2.64"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[tool.poetry.scripts]
app = "streamlit run src/app.py"

# Streamlitアプリのため、ビルドシステムは最小限に設定
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
